% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_engine.R
\name{add_data_language_engine}
\alias{add_data_language_engine}
\title{Activate data language engine for use in Rmarkdown documents}
\usage{
add_data_language_engine()
}
\description{
Activate data language engine for use in Rmarkdown documents
}
