% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{platform.newline}
\alias{platform.newline}
\title{Platform independent newline string}
\usage{
platform.newline(os = .Platform$OS.type)
}
\arguments{
\item{os}{Name of the operating system. Defaults to \code{.Platform$OS.type}}
}
\value{
For Windows, this should return \code{'\r\n'}, whereas for
other operating system it will return \code{'\n'}.
}
\description{
A simple function to determine the appropriate newline string for a given
operating system.
}
\author{
David M. Kaplan \email{dmkaplan2000@gmail.com}
}
