% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudio_addins.R
\encoding{UTF-8}
\name{remove_chunks_dialog}
\alias{remove_chunks_dialog}
\title{Invoke Rstudio addin to remove chunks from the active source document}
\usage{
remove_chunks_dialog(title = "Eliminate (data) chunks")
}
\arguments{
\item{title}{Text to place in title bar of gadget.}
}
\value{
Returns \code{TRUE} if one or more chunks were removed, \code{FALSE} otherwise.
}
\description{
The dialog will present a data table list of chunks in the source document. Select the rows
that correspond to the chunks that you wish to remove and hit the \code{Remove chunks} button
to remove them.
}
\details{
When the dialog is started, if the cursor is positioned inside a chunk in the source document,
then the row corresponding to this chunk will be selected by default.
}
\examples{
\dontrun{
remove_chunks_dialog()
}

}
\seealso{
Other Chunk tools: 
\code{\link{create_chunk}()},
\code{\link{create_data_chunk_dialog}()},
\code{\link{insert_data_chunk_dialog}()},
\code{\link{insert_data_chunk_template}()},
\code{\link{list_rmd_chunks}()}
}
\author{
David M. Kaplan \email{dmkaplan2000@gmail.com}
}
\concept{Chunk tools}
